requirements.txt
dataclasses-json==0.6.7
distro==1.9.0
jiter==0.5.0
langchain-community==0.2.12
marshmallow==3.22.0
mypy-extensions==1.0.0
openai==1.42.0
tqdm==4.66.5
typing-inspect==0.9.0
